Most reorg done, compiles now, no HA reporting yet or sensors
