+++ components/mitsubishi_itp/mitp_listener.h
@@ -1,13 +1,11 @@
 #pragma once
 
-#include "itp_packetprocessor.h"
-
 namespace esphome {
 namespace mitsubishi_itp {
 
 static constexpr char LISTENER_TAG[] = "mitsubishi_itp.listener";
 
-class MITPListener : public itp_packet::PacketProcessor {
+class MITPListener {
  public:
   virtual void publish() = 0;  // Publish only if the underlying state has changed
